Progress on collisions and UI
The hammer now collides with all objects that are tagged "enemy," plays a sound effect and smoke particle. The UI is now present in level 2 with new text fitting the new area, as well as a volume slider in all title and pause menus, as well as two different songs within the two levels.